requirements.txt
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
blis==0.7.8
catalogue==2.0.7
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cymem==2.0.6
Cython==0.29.28
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
gensim==4.2.0
h11==0.13.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
murmurhash==1.0.7
nest-asyncio==1.5.5
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathy==0.6.2
pickleshare==0.7.5
preshed==3.0.6
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pyaml==21.10.1
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
scipy==1.8.1
selenium==4.3.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.3
stack-data==0.3.0
thinc==8.1.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.10
wasabi==0.9.1
wcwidth==0.2.5
wincertstore==0.2
wsproto==1.1.0
